Setup 'npm test'
The old 'lint' task didn't work unless a portal was specified on the command line.
Change it so that if portal parameter is not specified, it will run on all
JS files, thus making it usable from 'npm test' which will be run from Jenkins.
If portal param is specified, it will on this portal JS files only (+ cross-portal
JS files).

Removed obsolete style options in JSHint that are deprecated and
better covered by JSCS.

Change-Id: Ic0e2d488bf39703e5b59e34c360a735b4cf954e3