Checks if localstorage is enabled and can be written to.
Prevents browsers from throwing error in private browsing mode.

This is a temporary solution. It does not guarantee that using
localStorage will not fail because all calls to the localStorage
API should be wrapped by try/catch. We will work on a patch ASAP.
However, it temporarily fixes private browsing in Safari, that is
why we still want to move forward with it.
Future patch: T129587

Bug: T129408
Change-Id: I949450a9a9aa94481e79e8ac0296618fc97142ad