Run purgeList maintenance script from /srv/mediawiki-staging
When configuration is altered in such a way to use former configuration could
lead to notices and warnings, /srv/mediawiki on the deployment server isn't
the optimal environment to run a maintenance script, as it will raise such
notices or warnings.

Yet, it's here convenient to run only one operation instead to need to login
to Terbium to run the purge script.

We so call it directly from /srv/mediawiki-staging, which is by definition
up to date.

Change-Id: Id3b1df08c2bb4546fafd5b6771de857ccbe703a8