Adding client-side translations for footer text.
This patch enables client-side translations for
all the content on the wikipedia portal page.

The translations for sister projects, footer text
and language button text have been available from
translatewiki for some time. Here these translations
are now utilized for the whole page and not just the
top ten links.

The dir attribute on the html tag has been removed
so that at least the text for rtl langs will be readable.

Bug: T136446
Change-Id: Ia3886baa8b5f81c119207d1e4720c12b44106320