Fix typeahead occasionally using wrong thumbnail and description
The 'sanitizedThumbURL' and 'descriptionText' variables can be carried
over from a previous iteration of the loop, if not explicitly set. Both
are now explicitly reset to their default values, ensuring that a
typeahead entry with a missing thumbnail or description cannot be given
the thumbnail and/or description of a previous entry.

Change-Id: I069854ffca7c48e19a219939c4ddb01c53f263b6