Fixing spacing in footer license
The default english version has a hard-coded full-stop after the
"Creative Commons Attribution-ShareAlike License." link, whereas some
translations don't have a full-stop, causing the license and "Terms
of Service" link to read as one.

This replaces the full-stop with a bullet character and adds some
padding before and after the bullet.

Bug: T178744
Change-Id: Iea25b07f20922eba0c36bb90e74bc4212c10f84b