SVG Sprites - Replacing PNG images with SVG sprite in the HTML
This patch replaces the images on the wikipedia.org portal with
an SVG sprite with PNG fallback instead.

Classes in the handlebars templates have been replaced to match the new
SVG sprite classes.

The new sprite.css file has been linked in index.handlebars, and the
sprite-template.handlebars file has been fixed to work with IE8.

The Gulp file has also been modified to run the svgSprite task during
`gulp watch`.

The Wikinews logo has been moved to the top-level img folder since it
and the Wikipedia logo are too big to be served as SVGs.

Bug:T129634, T177343
Change-Id: I78948556a830f722a3efad0f949d3ed01eb02a7e