Adding portals/deploy repo as submodule
Adds the `portals/deploy` repo as a submodule in place of the prod
folder. The prod folder must be deleted from GIT before the submodule
can be initialized in the next patch.

To better manage deployment assets, the prod folder is being moved into
a submodule so that it can be deployed as a seperate repository.

This lets us seperate changes to the source files (held in this repo),
from changes to the build (which are usually minor stats updates) but
change a large number of files due to the l10n updates.

This structure also makes managing deploys easier since each commit in
the `portals/deploy` repo represents a new build, and can be easily
rolled back to the previous build (which will always be the previous
commit) in case of an error.

NOTE: this patch does not initialize the `portals/deploy` submodule,
but it does remove the prod folder, so that it can be initialized in
the next patch.

Bug: T180777
Change-Id: I171f751288f7926535851801bc88e046b233b303