Updating header for mobile
Centering the wikipedia logo and header on mobile devices.
This patch reworks much of the header CSS and HTML to group
the Wikipedia wordmark and globe logo together. The globe logo
is now placed using aboslute positioning on desktop and
relative positioning on mobile.

For RTL languages, the logo is also now centered and positioned
to the right of the wordmark.

Bug: T184172
Change-Id: I58e4a9ee4644c15160d4bec4a4258933dc06540c