Catch exceptions on window.localStorage property access
If user blocks setting cookies, that automatically blocks the
usage of local storage and trying to access window.localStorage
throws an error.

Ibe75408913d4ffbf1791e1a6470296b29f809185 deals with the same
problem for mediawiki core, but changes in mw.storage were
not ported here.

Because of that, accessing list of all languages isn't working,
i.e. clicking on "Read Wikipedia in your language" does not
open the list.

Change-Id: I446730b20eeb5a3a53e7bdf5874ad4dc9eba9d1b