build: Fix writePurgeFile() compat with Node.js 10
The writePurgeFile() function was calling fs.writeFile() without
a callback. This was a mistake. It previously happened to work,
and would lead to the file asynchronously being written by Node
without Gulp waiting for it, with some race condition there.

On Node.js 10 this mistake is detected and results in an exception:

> TypeError [ERR_INVALID_CALLBACK]: Callback must be a function
>   at fs.writeFile (fs.js:1139:14)
>   at writePurgeFile (~/Git/portals/gulpfile.js:394:13)

Also, while at it, fix the following deprecation warning:

> (node:7179) [DEP0005] DeprecationWarning: Buffer() is deprecated due to
> security and usability issues. Please use the Buffer.from() instead.

However, rather than following the instruction here, pass the
contents as a string. There's no need for a buffer here.

Change-Id: Id08324f8ba8cf085119f167e930356e5d04a7cac