Update ESlint
Updates wikimedia-config-eslint and gulp-eslint to latest version.
This update requires syntax changes across a wide range of files,
however no functionality should be affected by this change.

With changes to wikimedia-config-eslint, much of the custom
.eslintrc.json can be removed, and separate eslintrc.json files
are added for folders running in Node & browser environments.

A comment capitalization rule and "root" rule have been added
to the main .eslintrc.json file.

An .eslintignore is also added to ignore prod & test dir.

Bug: T199825
Change-Id: I77fa8c9a9a39cbf7d94c92f555a5faa0bd85b200