Fix search suggestion clicking on Android Chrome.
On Chrome for Android, clicking search suggestion below the fold
causes the browser to scroll up back to the top of the page, to
make the search input visible, but the search suggestion link
is not followed.

Adds a method to manually change window.location when the search
input is blurred.

Bug: T221628
Change-Id: Ib96ce88ec953ffadb3db18a170093e0ff9fa3841