Fix symlink & CSS inlining issues with Gulp 4
A symlink named "portal" is required by our Apache config
to serve the portal static assets. This symlink is now created
via Gulp, which added a symlink method in version 4. Previously,
this symlink was added manually to the build asssets.

Replaces the "clean-css" minifier with cssnano due to some errors
when using it in conjunction with "gulp-inline".

Bug: T247996
Change-Id: I715fb7e5f82ef2dae0cab95cbca6e292204eb677