Add first English fundraising banner for portal
Please do not deploy before November 30.

Banner will be displayed if user:
- is in US, Canada, Australia, New Zealand, UK, or Ireland
- has English as primary language
- has not hidden banner on portal before
- has not hidden other fundraising banners (e.g. by donating)
and the year is still 2020

This first one is slightly different from previous setups: there is only one
banner and just the link to donatewiki is randomly chosen. One donatewiki page
has only the WMF logo as usual, the other has both the Wikipedia wordmark and
WMF logo.

Bug: T268359
Change-Id: If32896728c63d68897054bf98b7f51e68b406d3c