Remove PNG fallback workflow for SVG sprites
At this point all viable browsers support SVGs, so the PNG fallback
can be removed. This removes 4 outdated dependancies, one of which
required a C compiler (pngQuant) and helps improve the security
situation of the repo.

The following dependencies have been removed:

- gulp-replace
- gulp-svg2png
- pngquant-bin
- vinyl-paths

This brings down the security vulnerabilities via `npm audit` from 22 to 11.

Documentation has also been updated to reflect the fact that a PNG
fallback is no longer produced.

Bug: T258766
Bug: T262658
Change-Id: I40282b3b427ae7573fa6bf66620abe6df0065528