Remove exec() and add deleteFiles() function
At few places, exec() function is used which is a unix-centric shortcut
to execute certain commands. In portals, it is used to check the last
modified time of files and delete them if they exceed certain time limit.

To improve cross-OS compatibility, the function has been removed in this
patch and various methods of fileSystem module of Node has been used instead.

Bug: T274886
Change-Id: I454fb47b7d487e84709282dc603fc30953d16b8e