Followup 246c37b7, event-logging sends data in POST body not as URL params
Following up on "Migrate wikipedia.org analytics to Event Platform".
Previously the analytics sendbeacon encoded the event payload as
URL query parameters. For the Event Platform migration, the data has to
be sent in the request body (the second parameter in `navigator.sendbeacon()` ).

This patch ensures that happens and removes the img fallback for
event-logging since that method is not supported by the Event Platform.

Bug: T282012
Change-Id: I28c972238920d3726db51348371f93544d8a13bc