Override "skr-arab" to "skr" when building portals
The convention in MediaWiki is to use arabic (Urdu) script
for the Saraiki language by default.

To mimick this on the Wikimedia portals, we have to override
the translation file from "skr-arab.json" to "skr.json".

This patch makes that change during the build process by
adding a conditional for this language.

This is a rather inelegant solution, but it resolves the issue
of the language not appearing on the www.wiktionary.org portal.

Bug: T319137
Change-Id: Ib62b66861e864610b9749c8362f83b4145166687