Restore "Add <link rel="me"> to verify Wikipedia Mastodon account"
This reverts commit ed0c839a8e8414f1a80ed110735f4a5ec0ec751e.

After discussing with the WMF Comms team we clarified that it is
possible to verify multiple accounts to the same URL, which as far
as I know was the main concern (that a verified community account was
preventing a WMF-operated account) and haven't heard anything else.

Now that T324937 has landed, we could just verify it to e.g. the
English Wikipedia main page, but since we are aiming to be
multilingual it would be ideal to use the www.wikipedia.org portal.

Change-Id: I97dfd6c00bfaef316299ce214cd17a97a3ff7f8b