Next English fundraising test, including Fundraise Up in US
The bottom positioned banner is winning comfortably, so adopt that
(removing unneeded CSS but I may have missed some) and start next test.

For design, test is with and without the use of commons Montserrat and
Source Serif fonts. Note that to avoid redundancy in the html, this test
uses javascript to add the relevant class and adjust utm_source.

In the US only: test sending 50% of traffic to a Fundraise Up form,
and 50% to control form but with tweaks to closer match Fundraise Up
(lowest amount = $3, no monthly convert)

Change-Id: I17e4184b303f060b74d5582acfb177976b7a1cdd