Another English fundraising test, including Fundraise Up in US
With and without external fonts are even in donations, so adopt without
for better performance. Change design test to different font stacks.

In the US only: test sending 50% of traffic to a new Fundraise Up form
with a goal meter, and 50% to control form but with tweaks to closer
match Fundraise Up (lowest amount = $3, no pre-payment monthly convert
but add back post-payment)

Change-Id: I6b7344c070ac256243d386463ebbee200392e2de