Convert HTTP response to JSON Conditionally
Instead of converting the response body in httpGet()
to JSON on every request, first convert the response
to text, then try converting to JSON in a try/catch
statement, returning the text response if that fails.

Also, throw an error if one of the stats URLS
doesn't return a text response.

Bug: T359997
Change-Id: Iea2f16d412e09797dbdc36d1e659457dd8e1f7be