Fix number localization for added sites outside top 10
When a user has preferred browser languages outside the default top 10
sites, those languages will be added to the top sites by javascript.

In T283260 localized number formats were added for the numbers of pages.
However this only worked for the default top 10 sites, not any added
ones. Fix this by ensuring relevant locale is passed to the formatNumber
helper.

Bug: T369980
Change-Id: I4d405807d72a6451b5dd0a29a290b8de9f369858