Send correct family to search-redirect.php
search-redirect.php expects the second-level domain name in the `family`
query parameter, however, we’ve been passing the project name, which is
capitalized and could be translated (if I understand the code correctly,
it isn’t actually translated, but that’s just a performance optimization
and not a guarantee, should someone come up with a localization system
that doesn’t use JavaScript).

Since this second-level domain name also appears in the URL, change
`l10n.assets` to only pass the second-level domain name, and assemble
both the URL and the query parameter from that.

Bug: T391297
Change-Id: I8d17cdba33b05ba1631f80d59a6fd7b76e256d3a