Enable ESLint in data/ directory, lint fixes
ESLint was disabled in the entire data/ directory in
ac16a08736811b091c874b5fe870d17db00501cf, even though the only file that
doesn’t need to be linted is the auto-generated (and `.gitignore`d)
`site-stats.json`. So restrict the `.eslintignore` entry to that,
allowing spotting problems in all other files. Since the data/ directory
contains code that needs to be compatible with Node.js 18 instead of a
wide range of browsers, amend the configuration specific to that
directory so that constructs and library features such as optional catch
binding and native Promise are accepted by ESLint (per
https://compat-table.github.io/compat-table/es2016plus/,
Node 18 supports almost all features up to ES2021).

This uncovered a number of errors and warnings.

- Change the `capitalized-comments` lint configuration to allow
  consecutive comments – there was a violation, where simply an inline
  comment was continued in the next line; using a capital letter in the
  middle of the sentence would make no sense.
- Fix all other errors.
- Also fix a some (but not all) warnings, including one that’s outside
  of the data/ directory (so was already reported) but is easy to fix.

Change-Id: I22d65f257b82ff9f329ab4fd004a9a21a65f632f